requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
requests>=2.31.0
markdown>=3.4.0
premailer>=3.10.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
